initial dev branch commit
It should never be merged with the main library as it is purely for developmental purposes only. Add your own project files as you see fit for your preferred IDE but don't expect others to keep it up to date with the new files. Keep the main branch using the makefile for ease of use for end users